[ALPS06148961] Cast from enum to unsigned int to avoid negative value
Remove the risk of negative value indexing.

MTK-Commit-Id: e585fa15183b7be74ba9b261197ff938f78567d9

CR-Id: ALPS06148961
Feature: Wi-Fi Link Layer Statistics

+++ os/linux/gl_vendor.c
@@ -949,7 +949,7 @@ void dumpLinkStatsPeerInfo(struct STATS_LLS_PEER_INFO *peer, uint32_t idx)
 
 	DBGLOG(REQ, INFO, "Peer(%u) %u(%s)" MACSTR "%u %u %u [%u]",
 			idx,
-			peer->type, type[peer->type],
+			peer->type, type[(uint32_t)peer->type],
 			MAC2STR(peer->peer_mac_address),
 			peer->capabilities,
 			peer->bssload.sta_count,
